Update windows.md
In the "If Windows 11 setup is stuck due to lack of network connection:" section step 2 keyboard command seems to just change the volume on a Mac keyboard as F10 is shared with Volume Down. I was able to launch Command Prompt with fn + Shift + F10 which enables the function key instead of volume control.